Fixed two bugs concerning turn-point-detection in the rainflow counting module
1.	Consecutively duplicated values could lead to leaving out turns in the function `get_turns`.
•	If a turning point is directly following on consecutive duplicates it won't be considered as a turning point.
•	This is a result of the expression in former line 50, because the difference of consecutive values is 0 and though the multiplication with the next difference always results in 0.
•	To fix this, one must remove the duplicates before evaluating said expression, which could be done in two ways.
o	The easier way would be removing the duplicates in `_get_new_turns` before calling `get_turns`.
o	By doing so, the behaviour of `_get_new_turns` would be correct, but `get_turns` would still fail when confronted with consecutive duplicates as a standalone function.
o	That's why removing the duplicates inside of `get_turns` should be considered the better option.
o	To prevent a breaking change concerning the return value positions one must transform the positions in the non-duplicated samples back to the original samples.
o	This requires to invert the index shifting that occurs during removing an arbitrary amount of consecutive duplicates. (ll. 64-68)

2.	Samples without turns caused an `IndexError` in `_get_new_turns` (former line 86).
•	When get_turns doesn't detect turns, `turn_positions` is an empty ndarray.
•	Calling [-1] on an empty ndarray raises the `IndexError`.

Added tests for both cases.

Signed-off-by: Jakob Riebe (CR/AMP4) <fixed-term.Jakob.Riebe@de.bosch.com>