Improve bayesian test performance by introducing `test_bayesian_pseudo`
A new `test_bayesian_pseudo` test has been introduced that calls the pymc3
functions with smaller sample number and smaller tuning sample number. The
demanded tolerance is therefore decreased to 1e0.

Moreover the random_seed in for the bayes functions is now clamped in the tests
in order to avoid occasional failures.

Therefore the API of the `Bayesian` analyzer has been extended slightly in
order to accommodate the `tune` and `random_seed` parameters.

The old `test_bayesian` has been preserved in `test_bayesian_full` and been
marked with `slow_acceptance`. Tests with this marker are now disabled by
default in `pytest.ini`