Work around bug in pandas that puts NaNs into aligned Series.
We use pd.Series.align() to broadcast into a pd.Series or a pd.DataFrame when
both operands have common index levels.  That does not work when the second
operand is a pd.Series.

Workaround: We put the Series into a DataFrame, perform .align() and then take
the columns out of the DataFrame and have our aligned Series.

We can revert this once the bug in pandas is fixed.

pandas bug: https://github.com/pandas-dev/pandas/issues/46001

Signed-off-by: Johannes Mueller <johannes.mueller4@de.bosch.com>